engine: extract esm/layers/layer_manager.js
